backend/requirements.txt
fastapi
uvicorn[standard]
PyPDF2>=3.0.0
reportlab
pandas
openpyxl
SQLAlchemy
python-multipart
fastapi
PyPDF2>=3.0.0
reportlab
pandas
openpyxl
jinja2
passlib[bcrypt]
python-jose[cryptography]
python-dateutil
pdf2image
pillow
requests
